Add namespace for admin
