+++ app/views/admin/categories/edit.html.erb
@@ -6,7 +6,7 @@
             <h6>Update Category</h6>
         </div>
         <div class="form-padding">
-            <%= render 'categories/form' %>
+            <%= render 'form' %>
         </div>
     </div>
 </div>
