+++ app/views/admin/categories/index.html.erb
@@ -2,7 +2,7 @@
 <div class=container>
     <div class=row>
         <h1> All Categories </h1>
-        <%= link_to "Add Category", new_category_path, class: 'btn btn-primary' %>
+        <%= link_to "Add Category", new_admin_category_path, class: 'btn btn-primary' %>
         <%= will_paginate %>
         <table>
             <thead>
@@ -21,9 +21,9 @@
                     <td><%=  category.description %></td>
                     <td>0</td>
                     <td>
-                    <%= link_to "Edit", edit_category_path(category), 
+                    <%= link_to "Edit", edit_admin_category_path(category), 
                     class: 'btn blue lighten-1' %> 
-                    <%= link_to "Delete", category, method: :delete, 
+                    <%= link_to "Delete", [:admin, category], method: :delete, 
                     data: {confirm: "Are you sure you want to delete this category?"},
                     class: 'btn red darken-1' %>
                     </td>
