+++ app/views/admin/categories/new.html.erb
@@ -6,7 +6,7 @@
             <h6>Add Category</h6>
         </div>
         <div class="form-padding">
-            <%= render 'categories/form' %>
+            <%= render 'form' %>
         </div>
     </div>
 </div>
