+++ app/views/layouts/_header.html.erb
@@ -9,8 +9,7 @@
                 <li><%= link_to "Users", users_path %></li>
                 <ul id="dropdown1" class="dropdown-content">
                     <% if current_user.admin? %>
-                         <li><%= link_to "Categories", categories_path %></li>
-                    <% else %>
+                         <li><%= link_to "Categories", admin_categories_path %></li>
                     <% end %>
                     <li><%= link_to "Profile", current_user %></li>
                     <li><%= link_to "Settings", edit_user_path(current_user) %></li>
