+++ config/routes.rb
@@ -1,11 +1,5 @@
 Rails.application.routes.draw do
 
-  get 'categories/index'
-  get 'categories/show'
-  get 'categories/new'
-  get 'categories/create'
-  get 'categories/edit'
-  get 'categories/update'
   root 'static_pages#home'
   get  '/about',   to: 'static_pages#about'
   get  '/contact', to: 'static_pages#contact'
@@ -20,5 +14,8 @@ Rails.application.routes.draw do
     end
   end
   resources :relationships, only: [:create, :destroy]
-  resources :categories
+
+  namespace :admin do
+    resources :categories
+  end
 end
